[policy in code] Add support for AZ, scheduler and message resource [9/10]
This patch adds policy in code support for availability_zone,
scheduler_stats and message resources.

Partial-Implements: blueprint policy-in-code

+++ devstack/plugin.sh
@@ -145,7 +145,9 @@ function configure_manila {
     fi
     sudo chown $STACK_USER $MANILA_CONF_DIR
 
-    cp -p $MANILA_DIR/etc/manila/policy.json $MANILA_CONF_DIR
+    if [[ -f $MANILA_DIR/etc/manila/policy.json ]]; then
+        cp -p $MANILA_DIR/etc/manila/policy.json $MANILA_CONF_DIR
+    fi
 
     # Set the paths of certain binaries
     MANILA_ROOTWRAP=$(get_rootwrap_location manila)
